requirements_core.txt
# requirements_core.txt

websocket-client>=1.3.0

websockets>=10.0

quart>=0.20.0

jsonschema>=4.25.0

numpy>=2.3.2

scipy>=1.16.0

pyserial>=3.5